requirements.txt
# CodeCollab Requirements
# Python 3.12+ required

# === CORE DEPENDENCIES ===

# Strands Multi-Agent Framework (AWS SDK - always use latest)
strands-agents
strands-agents-tools

# AWS Bedrock (for Claude Sonnet 4)
boto3>=1.40.0
botocore>=1.40.0

# Environment Variables
python-dotenv==1.1.1

# === OPTIONAL DEVELOPMENT DEPENDENCIES ===

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code Quality (optional)
black>=23.0.0
flake8>=7.0.0

# Note: No web framework dependencies needed
# - simple_api_server.py uses Python stdlib only (http.server)
# - No FastAPI, uvicorn, or websockets required
# - GitHub tools use subprocess and git commands (no PyGithub needed)

# System Requirements:
# - git command-line tool (for GitHub repository cloning)
# - AWS credentials configured (for Bedrock access)
